process of weeding out and condensing project
