+++ app/controllers/details.js
@@ -2,7 +2,4 @@ import Ember from 'ember';
 
 export default Ember.Controller.extend({
 
-// id = this.get('model.source-config');
-// label = this.get('model.source-config');
-
 });
