+++ app/controllers/status/source-detail/index.js
@@ -1,4 +1,5 @@
 import Ember from 'ember';
 
 export default Ember.Controller.extend({
+
 });
