+++ app/controllers/status/source-detail/log-detail.js
