--- app/routes/source-config.js
@@ -1,10 +1,0 @@
-import Ember from 'ember';
-
-export default Ember.Route.extend({
- model(){
-	 return Ember.RSVP.hash({
-	 harvestlog: this.store.findAll('harvest-log'),
- 	 sourceconfig: this.store.findAll('source-config'),
- 	});
- }
-});
