--- app/routes/source.js
@@ -1,12 +1,0 @@
-import Ember from 'ember';
-
-export default Ember.Route.extend({
-
-	model() {
-		return Ember.RSVP.hash({
-			harvestlog: this.store.findAll('harvest-log'),
-			sourceconfig: this.store.findAll('source-config'),
-			source: this.store.findAll('source'),
-		});
-	}
-});
