--- app/routes/status-board.js
@@ -1,12 +1,0 @@
-import Ember from 'ember';
-
-export default Ember.Route.extend({
-	queryParams: {
-	    page: {
-	      refreshModel: true
-	    }
-	  },
-	model(params) {
-		return this.get('store').query('source-config', params);
-	},
-});
