--- app/templates/source.hbs
@@ -1,47 +1,0 @@
-{{outlet}}
-
-<h2> Harvest log Data</h2>
-{{#each model as |harvest-log|}}
-
-	{{harvest-log.status}}<br>
-	<br>
- 	{{harvest-log.completions}}<br>
-{{/each}}
-	<div onload= "loadPage" class="masthead">
-		<h1 align="center" class="text-muted" style= "font-size:5rem; font-family:courier;">HARVEST LOG</h1>
-	</div>
-
-<div>
-<style>
-table, th, td {
-		padding: 3rem;
-		padding-top: 2rem;
-		border-bottom: solid #d9d9d9;
-}
-th {
-	font-family:Courier;
-	font-size: 3rem;
-	color: #0099ff;
-	padding-bottom: 2px;
-}
-td{
-	font-family:courier;
-	font-size: 2rem;
-
-}
-
-</style>
-	<table style="width:95%" align="center" class="table">
-		{{#each sources as |source index|}}
-    {{log source}}
-	  <tr>
-      <td>{{source.attributes.longTitle}}</td>
-      <td><img height="30" width="30" style="align:center" src=assets/images/created.png /></td>
-      <td><img height="30" width="30" style="align:center" src=assets/images/fail.png /> </td>
-      <td><img height="30" width="30" style="align:center" src=assets/images/succeed.png /> </td>
-      <td><img height="30" width="30" style="align:center" src=assets/images/succeed.png /> </td>
-      <td><img height="30" width="30" style="align:center" src=assets/images/fail.png /> </td>
-    </tr>
-    {{/each}}
-      </table>
-    </div>
