--- app/templates/status-board.hbs
@@ -1,43 +1,0 @@
-{{outlet}}
-
-{{page-nav atLastPage=atLastPage atFirstPage=atFirstPage nextPage=(action 'nextPage') prevPage=(action 'prevPage')}}
-
-<h1 align="center" class="text-muted" style="font-size:5rm; font-family:courier;">{{model.label}}</h1>
-
-<div>
-	<style>
-	table, th, td {
-		font-family: courier;
-		padding: 3rem;
-		padding-top: 2rem;
-	}
-	th {
-		padding: 3rem;
-		font-family:Courier;
-		font-size: 2rem;
-		color: #0099ff;
-	}
-	td{
-		padding: 3rem;
-		font-family:courier;
-		font-size: 2rem;
-	}
-	</style>
-		<table style="width:95%" align="center" class="table">
-			<tr>
-				<th>SOURCE</th>
-				<th>RECENT HARVESTS</th>
-			</tr>
-
-<tbody>
-{{#each model as |sc|}}
-		<tr>
-			<td>{{#link-to "harvestloglist" sc.label sc.id}}{{#status-board title=sc.label}}{{/status-board}}{{/link-to}}</td>
-			<td>{{#source-health sourceConfigId=sc.id}}
-			{{/source-health}}</td>
-	</tr>
-{{/each}}
-</tbody>
-
-</table>
-</div>
